pyproject.toml
[tool.poetry]
name = "trending-repository-summarizer"
version = "0.1.0"
description = "Summarize trending repositories on GitHub."
authors = ["Ryohei Machida <r-machida@analytics-jp.com>"]
packages = [{ include = "*", from = "src" }]
readme = "README.md"
homepage = "https://github.com/smarlhens/python-boilerplate#readme"
repository = "https://github.com/smarlhens/python-boilerplate"

[tool.poetry.scripts]
main = "trending_repository_summarizer.main:main"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
pydantic = "^2.10.5"
lxml = "^5.3.0"
langchain-openai = "^0.3.0"
markdown = "^3.7"
jinja2 = "^3.1.5"
universal-pathlib = "^0.2.6"
fsspec = {extras = ["s3"], version = "^2024.12.0"}
notion-client = "^2.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "8.2.2"
pytest-cov = "5.0.0"
mypy = "1.10.0"
bandit = "1.7.9"
docformatter = "1.7.5"
ruff = "0.4.9"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
backend-path = ["src"]

[tool.pytest.ini_options]
addopts = "-vvv"
testpaths = "tests"

[tool.ruff]
extend-exclude = ["__pycache__", "build", "dist"]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
extend-select = [
    "C4",
    "D200",
    "D201",
    "D204",
    "D205",
    "D206",
    "D210",
    "D211",
    "D213",
    "D300",
    "D400",
    "D402",
    "D403",
    "D404",
    "D419",
    "E",
    "F",
    "G010",
    "I001",
    "INP001",
    "N805",
    "PERF101",
    "PERF102",
    "PERF401",
    "PERF402",
    "PGH004",
    "PGH005",
    "PIE794",
    "PIE796",
    "PIE807",
    "PIE810",
    "RET502",
    "RET503",
    "RET504",
    "RET505",
    "RUF015",
    "RUF100",
    "S101",
    "T20",
    "W",
]

[tool.mypy]
files = ["src", "tests"]
strict = true
